Get rid of the BridgeSupportTest hierarchy, move those methods to util classes, and make test classes independent.
